requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
matplotlib>=3.7
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
pillow>=10.0
python-multipart>=0.0.9
